ServiceInstance takes float instead of int for gbRamUsed and coresUsed

+++ object_database/service_manager/ServiceInstance.py
@@ -56,8 +56,8 @@ class Service:
     service_class_name = str
 
     # per service, how many do we use?
-    gbRamUsed = int
-    coresUsed = int
+    gbRamUsed = float
+    coresUsed = float
     validPlacementGroups = TupleOf(str)
     isSingleton = bool
